// src/types/game.d.ts
//adaped from https://github.com/wabbajack-tools/wabbajack/blob/master/Wabbajack.Common/GameMetaData.cs

export type Game =
  | 'morrowind'
  | 'oblivion'
  | 'fallout3'
  | 'falloutnewvegas'
  | 'skyrim'
  | 'enderal'
  | 'skyrimspecialedition'
  | 'fallout4'
  | 'skyrimvr'
  | 'fallout4vr'
  | 'darkestdungeon'
  | 'witcher3'
  | 'stardewvalley';